+++ src/tabpfn/base.py
@@ -223,7 +223,7 @@ def create_inference_engine(  # noqa: PLR0913
     fit_mode: Literal["low_memory", "fit_preprocessors", "fit_with_cache", "batched"],
     devices_: Sequence[torch.device],
     rng: np.random.Generator,
-    n_jobs: int,
+    n_preprocessing_jobs: int,
     byte_size: int,
     forced_inference_dtype_: torch.dtype | None,
     memory_saving_mode: bool | Literal["auto"] | float | int,
@@ -246,7 +246,8 @@ def create_inference_engine(  # noqa: PLR0913
         fit_mode: Determines how we prepare inference (pre-cache or not).
         devices_: The devices for inference.
         rng: Numpy random generator.
-        n_jobs: Number of parallel CPU workers.
+        n_preprocessing_jobs: Number of parallel CPU workers to use for the
+            preprocessing.
         byte_size: Byte size for the chosen inference precision.
         forced_inference_dtype_: If not None, the forced dtype for inference.
         memory_saving_mode: GPU/CPU memory saving settings.
@@ -268,7 +269,7 @@ def create_inference_engine(  # noqa: PLR0913
             ensemble_configs=ensemble_configs,
             rng=rng,
             model=model,
-            n_workers=n_jobs,
+            n_preprocessing_jobs=n_preprocessing_jobs,
             dtype_byte_size=byte_size,
             force_inference_dtype=forced_inference_dtype_,
             save_peak_mem=memory_saving_mode,
@@ -279,7 +280,7 @@ def create_inference_engine(  # noqa: PLR0913
             y_train=y_train,
             cat_ix=cat_ix,
             ensemble_configs=ensemble_configs,
-            n_workers=n_jobs,
+            n_preprocessing_jobs=n_preprocessing_jobs,
             model=model,
             rng=rng,
             dtype_byte_size=byte_size,
@@ -294,7 +295,7 @@ def create_inference_engine(  # noqa: PLR0913
             cat_ix=cat_ix,
             model=model,
             ensemble_configs=ensemble_configs,
-            n_workers=n_jobs,
+            n_preprocessing_jobs=n_preprocessing_jobs,
             devices=devices_,
             dtype_byte_size=byte_size,
             rng=rng,